+++ lib/src/reactives.dart
@@ -16,19 +16,19 @@ class _ReactivesBuilderState<T extends Reactive> extends State<ReactivesBuilder>
 
   void initState() {
     super.initState();
-    widget.states.forEach((f) => f.addListener(_onChange));
+    widget.states.forEach((f) => f.listen(_onChange));
   }
 
   void dispose() {
-    widget.states.forEach((f) => f.removeListener(_onChange));
+    widget.states.forEach((f) => f.forget(_onChange));
     super.dispose();
   }
 
   void didUpdateWidget(ReactivesBuilder oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (widget.states != oldWidget.states) {
-      oldWidget.states.forEach((f) => f.removeListener(_onChange));
-      widget.states.forEach((f) => f.addListener(_onChange));
+      oldWidget.states.forEach((f) => f.forget(_onChange));
+      widget.states.forEach((f) => f.listen(_onChange));
     }
   }